Refactor: Split CareerDashboard component
Refactor CareerDashboard component into smaller files for better organization and maintainability.
