+++ src/components/admin/careers/types.ts
@@ -6,5 +6,12 @@ export interface CareerPost {
   type: string;
   description: string;
   requirements: string;
+  qualifications?: string;
+  benefits?: string;
+  salary?: string;
+  applicationProcess?: string;
   date: string;
+  status: 'active' | 'draft' | 'closed';
 }
+
+export type ViewMode = 'list' | 'editor';
